Don't separate data and stores.
