+++ assets/js/googlesitekit/data/store.js
@@ -5,7 +5,7 @@
 /**
  * Internal dependencies
  */
-import siteStore, { STORE_NAME as siteStoreName } from 'assets/js/googlesitekit/site';
+import siteStore, { STORE_NAME as siteStoreName } from 'assets/js/googlesitekit/data/stores/site';
 
 export const registerSiteKitStores = ( registry ) => {
 	registry.registerStore( siteStoreName, siteStore );
