+++ assets/js/googlesitekit/data/stores/site/reset.js
